fix MapSectionManager : null일 때 예외처리, refactor createBitmapWithColor
